refactor: change filenames from camel case to kebab case in products list folder
