+++ apps/core/app/dashboard/products/_components/table/product-table-head/product-table-head.tsx
