+++ apps/core/app/dashboard/products/_components/table/product-table-pages/product-table-pages.tsx
