+++ apps/core/app/dashboard/products/_components/tabs/product-tabs.tsx
