+++ apps/core/app/dashboard/products/page.tsx
@@ -3,8 +3,8 @@ import CirclePlusIcon from "@repo/icons/circle-plus";
 import { Button } from "@repo/ui/components/button";
 import Typography from "@repo/ui/components/typography";
 import { useState } from "react";
-import { ProductTabs } from "./_components/tabs/productTabs";
-import { ProductTable } from "./_components/table/ProductTableHead";
+import { ProductTabs } from "./_components/tabs/product-tabs";
+import { ProductTable } from "./_components/table/product-table-head";
 
 const tabs = [
   { value: "all", label: "Products" },
